[rename] search 페이지 관련 커스텀훅 폴더로 묶음
