+++ hooks/search/useClearProductFilter.ts
